MSLMR_TRDNG/requirements.txt
# Científica y análisis de datos
numpy==1.24.3
pandas==2.0.3
scipy==1.10.1

# Machine Learning
scikit-learn==1.2.2
tensorflow==2.12.0
keras==2.12.0
lightgbm==3.3.5
xgboost==1.7.5

# Base de datos y ORM
SQLAlchemy==2.0.20
psycopg2-binary==2.9.7
redis==4.5.4
pymongo==4.3.3

# Trading
ccxt==4.0.0
ta-lib==0.4.20

# Web y API
flask==2.3.2
flask-restful==0.3.9
requests==2.30.0

# Logging y monitoreo
structlog==23.1.0
python-json-logger==2.0.7
prometheus-client==0.16.0

# Visualización
plotly==5.14.1
matplotlib==3.7.1

# Configuración y utilidades
python-dotenv==1.0.0
pyyaml==6.0

# Desarrollo y testing
pytest==7.3.1
pytest-cov==4.0.0
mypy==1.3.0
black==23.3.0
flake8==6.0.0

# Seguridad
cryptography==40.0.2

# Documentación
sphinx==6.2.1

# Misc
tqdm==4.65.0
